Split API controller
